extracted category management to a global context. Category Picker now uses the global context instead of managing them itself.
